add day 7 part 2 solution

+++ src/Day07/Equation.cs
@@ -3,6 +3,6 @@
     public class Equation
     {
         public long TestValue { get; set; }
-        public List<int> Numbers { get; set; } = [];
+        public List<long> Numbers { get; set; } = [];
     }
 }
